Title: 최소 에너지 비용, Time: 263ms, Memory: 16MB, Status: Wrong Answer - CodeTree

+++ 240429/최소 에너지 비용/minimum-energy-cost.kt
@@ -21,9 +21,9 @@ fun main() {
         L[i] = min(L[i - 1], cost[i])
     }
 
-    val result = IntArray(n) // result[i]: i-1에서 i로 이동할 때, '에너지 x 최소 비용'
+    val result = LongArray(n) // result[i]: i-1에서 i로 이동할 때, '에너지 x 최소 비용'
     for (i in 1..n-1) {
-        result[i] = L[i - 1] * energy[i - 1]
+        result[i] = (L[i - 1] * energy[i - 1]).toLong()
     }
 
     var answer = 0L
